requirements.txt
# Backend API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
pyyaml==6.0.1
pandas==2.1.4
